# backend/openapi_snapshot.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "My Darrin API",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/v1/catalog/domains":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "Catalog"
                                                                       ],
                                                              "summary":  "List Domains",
                                                              "operationId":  "list_domains_api_v1_catalog_domains_get",
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Successful Response",
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "items":  {
                                                                                                                                                                  "$ref":  "#/components/schemas/DomainResponse"
                                                                                                                                                              },
                                                                                                                                                    "type":  "array",
                                                                                                                                                    "title":  "Response List Domains Api V1 Catalog Domains Get"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          },
                                                  "post":  {
                                                               "tags":  [
                                                                            "Catalog"
                                                                        ],
                                                               "summary":  "Create New Domain",
                                                               "operationId":  "create_new_domain_api_v1_catalog_domains_post",
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/DomainCreate"
                                                                                                                                       }
                                                                                                                        }
                                                                                               },
                                                                                   "required":  true
                                                                               },
                                                               "responses":  {
                                                                                 "201":  {
                                                                                             "description":  "Successful Response",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/DomainResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "422":  {
                                                                                             "description":  "Validation Error",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         }
                                                                             }
                                                           }
                                              },
...
                  "/api/v1/catalog/categories":  {
                                                     "get":  {
                                                                 "tags":  [
                                                                              "Catalog"
                                                                          ],
                                                                 "summary":  "List Categories",
                                                                 "operationId":  "list_categories_api_v1_catalog_categories_get",
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "items":  {
                                                                                                                                                                     "$ref":  "#/components/schemas/CategoryResponse"
                                                                                                                                                                 },
                                                                                                                                                       "type":  "array",
                                                                                                                                                       "title":  "Response List Categories Api V1 Catalog Categories Get"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             },
                                                     "post":  {
                                                                  "tags":  [
                                                                               "Catalog"
                                                                           ],
                                                                  "summary":  "Create New Category",
                                                                  "operationId":  "create_new_category_api_v1_catalog_categories_post",
                                                                  "requestBody":  {
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/CategoryCreate"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  },
                                                                                      "required":  true
                                                                                  },
                                                                  "responses":  {
                                                                                    "201":  {
                                                                                                "description":  "Successful Response",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/CategoryResponse"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            },
                                                                                    "422":  {
                                                                                                "description":  "Validation Error",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            }
                                                                                }
                                                              }
                                                 },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CategoryResponse":  {
                                                                "properties":  {
                                                                                   "domain_id":  {
                                                                                                     "type":  "integer",
                                                                                                     "title":  "Domain Id"
                                                                                                 },
                                                                                   "name":  {
                                                                                                "type":  "string",
                                                                                                "title":  "Name"
                                                                                            },
                                                                                   "slug":  {
                                                                                                "type":  "string",
                                                                                                "title":  "Slug"
                                                                                            },
                                                                                   "is_active":  {
                                                                                                     "type":  "boolean",
                                                                                                     "title":  "Is Active",
                                                                                                     "default":  true
                                                                                                 },
                                                                                   "id":  {
                                                                                              "type":  "integer",
                                                                                              "title":  "Id"
                                                                                          }
                                                                               },
                                                                "type":  "object",
                                                                "required":  [
                                                                                 "domain_id",
                                                                                 "name",
                                                                                 "slug",
                                                                                 "id"
                                                                             ],
                                                                "title":  "CategoryResponse"
                                                            },
...
                                       "DomainCreate":  {
                                                            "properties":  {
                                                                               "name":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Name"
                                                                                        },
                                                                               "slug":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Slug"
                                                                                        },
                                                                               "is_active":  {
                                                                                                 "type":  "boolean",
                                                                                                 "title":  "Is Active",
                                                                                                 "default":  true
                                                                                             }
                                                                           },
                                                            "type":  "object",
                                                            "required":  [
                                                                             "name",
                                                                             "slug"
                                                                         ],
                                                            "title":  "DomainCreate"
                                                        },
                                       "DomainResponse":  {
                                                              "properties":  {
                                                                                 "name":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Name"
                                                                                          },
                                                                                 "slug":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Slug"
                                                                                          },
                                                                                 "is_active":  {
                                                                                                   "type":  "boolean",
                                                                                                   "title":  "Is Active",
                                                                                                   "default":  true
                                                                                               },
                                                                                 "id":  {
                                                                                            "type":  "integer",
                                                                                            "title":  "Id"
                                                                                        }
                                                                             },
                                                              "type":  "object",
                                                              "required":  [
                                                                               "name",
                                                                               "slug",
                                                                               "id"
                                                                           ],
                                                              "title":  "DomainResponse"
                                                          },
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                   },
                       "securitySchemes":  {
                                               "HTTPBearer":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              }
                                           }
                   }
}